### 2024-10-23 #7401 @warrickball
In the meantime, there are a few things you can have a quick look at.

* I see Marla Geha is included as an author but they aren't obviously connected to the software. I don't doubt that they contributed but could you briefly comment on their contribution? You may also want to mention them as a contributor in the docs. They don't appear there either.
* Please add the county to your affiliation. ("USA" is fine.)
* I suspect you're selecting references from a much larger BibTeX file. Could you trim the BibTeX file down to the minimum so we only check things like DOIs for the references actually used in the paper?
* Along that vein, note that a lot of references don't have DOIs. That's fine—some references don't—but if you could do a sweep (after reducing the BibTeX list and running `@editorialbot check references` again) to see if you can fill in any gaps that'd be great. If you're getting your data from NASA ADS, it should be up to date.
