+++ README.MD
@@ -1,3 +1,11 @@
+<!-- Joss badge -->
+HTML: <a href="https://joss.theoj.org/papers/a26f20391c1360ed1f2806fa09065816"><img src="https://joss.theoj.org/papers/a26f20391c1360ed1f2806fa09065816/status.svg"></a>
+Markdown: [![status](https://joss.theoj.org/papers/a26f20391c1360ed1f2806fa09065816/status.svg)](https://joss.theoj.org/papers/a26f20391c1360ed1f2806fa09065816)
+
+<!-- MIT Licence badge -->
+HTML: <a href="https://opensource.org/licenses/MIT"><img src="https://img.shields.io/badge/License-MIT-yellow.svg"></a>
+Markdown: [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+
 Dynamicall is a package written to fit kinematic observations of Dwarf Galaxies using Bayesian Inference.
 It is written in python using [Jax](https://jax.readthedocs.io/en/latest/) and the inference is done using [Numpyro](https://num.pyro.ai/en/stable/#).
 
