### 2024-11-19 769f18e
trimmed down bibfile
