+++ paper.bib
@@ -1,1453 +1,29 @@
 @article{emcee,
-   author = {{Foreman-Mackey}, D. and {Hogg}, D.~W. and {Lang}, D. and {Goodman}, J.},
-    title = {emcee: The MCMC Hammer},
+  author = {{Foreman-Mackey}, D. and {Hogg}, D.~W. and {Lang}, D. and {Goodman}, J.},
+  title = {emcee: The MCMC Hammer},
   journal = {PASP},
-     year = 2013,
-   volume = 125,
-    pages = {306-312},
-   eprint = {1202.3665},
-      doi = {10.1086/670067}
-}
-
-
+  year = 2013,
+  volume = 125,
+  pages = {306-312},
+  eprint = {1202.3665},
+  doi = {10.1086/670067}}
+  
 @ARTICLE{Vasiliev2019-aa,
-       author = {{Vasiliev}, Eugene},
-        title = "{AGAMA: action-based galaxy modelling architecture}",
-      journal = {\mnras},
-     keywords = {methods: numerical, galaxies: kinematics and dynamics, Astrophysics - Astrophysics of Galaxies},
-         year = 2019,
-        month = jan,
-       volume = {482},
-       number = {2},
-        pages = {1525-1544},
-          doi = {10.1093/mnras/sty2672},
-archivePrefix = {arXiv},
-       eprint = {1802.08239},
- primaryClass = {astro-ph.GA},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2019MNRAS.482.1525V},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-@ARTICLE{Governato2015,
-  author = {{Governato}, F. and {Weisz}, D. and {Pontzen}, A. and {Loebman}, S. and {Reed}, D. and {Brooks}, A.~M. and {Behroozi}, P. and {Christensen}, C. and {Madau}, P. and {Mayer}, L. and {Shen}, S. and {Walker}, M. and {Quinn}, T. and {Keller}, B.~W. and {Wadsley}, J.},
-  title = "{Faint dwarfs as a test of DM models: WDM versus CDM}",
+  author = {{Vasiliev}, Eugene},
+  title = "{AGAMA: action-based galaxy modelling architecture}",
   journal = {\mnras},
-  keywords = {hydrodynamics, Galaxy: formation, galaxies: formation, Astrophysics - Astrophysics of Galaxies, Astrophysics - Cosmology and Nongalactic Astrophysics},
-  year = 2015,
-  month = mar,
-  volume = {448},
-  number = {1},
-  pages = {792-803},
-  doi = {10.1093/mnras/stu2720},
-  archivePrefix = {arXiv},
-  eprint = {1407.0022},
-  primaryClass = {astro-ph.GA},
-  adsurl = {https://ui.adsabs.harvard.edu/abs/2015MNRAS.448..792G},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}}
-
-@ARTICLE{Acquaviva2012-xu,
-  title = "{SURVEY DESIGN FOR SPECTRAL ENERGY DISTRIBUTION FITTING: A FISHER MATRIX APPROACH}",
-  author = "Acquaviva, Viviana and Gawiser, Eric and Bickerton, Steven J and Grogin, Norman A and Guo, Yicheng and Lee, Seong-Kook",
-  journal = "ApJ",
-  volume =  749,
-  number =  1,
-  pages = "72",
-  month =  apr,
[…]
