### 2025-02-12 #7401 @aneeshnaik
Hi @JBorrow, unfortunately I'm very busy at the moment as I've just started a new job, so I don't think I can commit the time right now. Very sorry about this, looks like a very interesting paper/package!

@michael-petersen is also a JOSS reviewer, I wonder if he might be interested and available?
