### 2025-02-19 #7794 @editorialbot
[REVIEW]: DynamicAll
**Submitting author:** @jguerra-astro (<a href="http://orcid.org/0000-0002-7600-5110">Juan Guerra</a>)
**Repository:** https://github.com/jguerra-astro/dynamicall
**Branch with paper.md** (empty if default branch): main
**Version:** e.g V1.0
