### 2025-05-02 jguerra-astro/dynamicall#4 @sophialilleengen
[JOSS REVIEW] Installation issues
This is part of the [JOSS review 7794](https://github.com/openjournals/joss-reviews/issues/7794). I followed the installation instructions in the docs using conda. During the install process, several issues arised for me:

- [ ] In instructions, `# or whatever you want to call your environment` messes with installation in terminal. Should be removed.
- [ ]  conda install does not find `astroquery`, `numpyro`; `-c conda-forge` works. Installing arviz this way gives the wrong version (0.17 vs 0.20). Please test and update the instructions
- [ ] The wrong arviz version leads to a fail when testing the installation for stable scipy version '1.15.1’. Please change the instructions to use conda-forge.

I could fix these three and install dynamicAll. However, when testing the installation with `import dynamicAll`, I get an ImportError for Agama. This might be due to the set python version in the install instructions (`conda create -n dynamicall python=3.10 `) for which Agama was already installed.

```Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/slill/dynamicall/src/dynamicAll/__init__.py", line 3, in <module>
    from . import models
  File "/Users/slill/dynamicall/src/dynamicAll/models.py", line 19, in <module>
    from .base import JaxDensity
  File "/Users/slill/dynamicall/src/dynamicAll/base.py", line 12, in <module>
    import agama
  File "/Users/slill/.local/lib/python3.10/site-packages/agama/__init__.py", line 1, in <module>
    from .agama import *               # import everything from the C++ library
ImportError: dlopen(/Users/slill/.local/lib/python3.10/site-packages/agama/agama.so, 0x0002): tried: '/Users/slill/.local/lib/python3.10/site-packages/agama/agama.so' (slice is not valid mach-o file), '/System/Volumes/Preboot/Cryptexes/OS/Users/slill/.local/lib/python3.10/site-packages/agama/agama.so' (no such file), '/Users/slill/.local/lib/python3.10/site-packages/agama/agama.so' (slice is not valid mach-o file)
```
