### 2025-05-02 jguerra-astro/dynamicall#5 @sophialilleengen
[JOSS REVIEW] Remove custom stylefile
This is part of the [JOSS review 7794](https://github.com/openjournals/joss-reviews/issues/7794). Most notebooks still have the custom matplotlib style `plt.style.use("/Users/juan/phd/matplotlib/jdefault.mplstyle")`. This should be removed.
