### 2025-05-02 jguerra-astro/dynamicall#6 @sophialilleengen
[JOSS REVIEW] Notebooks missing data and errors
This is part of [JOSS review 7794](https://github.com/openjournals/joss-reviews/issues/7794).

Several files refer to data saved locally. Please provide the data in the repository and update the links. The notebooks are [data.py](https://github.com/jguerra-astro/dynamicall/blob/main/src/dynamicAll/data.py), [example.ipynb](https://github.com/jguerra-astro/dynamicall/blob/main/notebooks/example.ipynb), [tutorial_data.ipynb](https://github.com/jguerra-astro/dynamicall/blob/main/docs/notebooks/tutorial_data.ipynb), [tutorial_draco_NFW.ipynb](https://github.com/jguerra-astro/dynamicall/blob/main/docs/notebooks/tutorial_draco_NFW.ipynb), [tutorial_dracoish.ipynb](https://github.com/jguerra-astro/dynamicall/blob/main/docs/notebooks/tutorial_dracoish.ipynb), [tutorial_dsphsim.ipynb](https://github.com/jguerra-astro/dynamicall/blob/main/docs/notebooks/tutorial_dsphsim.ipynb).

Many of the notebooks in this repository show errors and are not run in order. It would be good to rerun all of them and fix any errors that show up.
