### 2025-05-02 jguerra-astro/dynamicall#7 @sophialilleengen
[JOSS REVIEW] Community guidelines are missing
This is part of the [JOSS review 7794](https://github.com/openjournals/joss-reviews/issues/7794).

As part of the checklist, there should be community guidelines. Please add these.

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
