### 2025-05-23 #7794 @JBorrow
Hi @jguerra-astro, just checking in to make sure you recieved the issues raised by @sophialilleengen and are able to start making progress towards resolving them?
