### 2025-10-31 #7794 @jguerra-astro
Apologies for the delay on this. I will begin working on the issues now. Thank you for your patience!
