### 2025-12-12 #7794 @JBorrow
Hi @jguerra-astro just checking in to see if you've been able to address any of the outstanding issues and to potentially get a timeline from you?
